requirements.txt
dash
plotly
pandas
psycopg-binary
sqlalchemy
python-dotenv